pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "Taskbeard"
version = "0.5.0"
description = "Simple project-management for FRC teams"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"click==8.2.1",
	"fastapi==0.116.1",
	"google-auth>=2.38.0",
	"pyjwt>=2.10.0",
	"pytest==8.4.1",
	"pyyaml==6.0.2",
	"requests>=2.32.0",
	"uvicorn[standard]==0.35.0",
	"watchfiles==1.1.0",
]

[project.scripts]
taskbeard = "main:cli"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.setuptools]
py-modules = ["main"]
